TASK#16: Refactoring 1. Moves App Component into "component" dir
