+++ project/src/components/app/app.tsx
@@ -1,21 +1,21 @@
 import React, { useEffect } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { Route, Switch } from 'react-router-dom';
-import { AppRoute } from './constants';
-import { initializeApp } from './store/app/app-thunks';
-import { AsyncDispatch, State } from './store/store';
-import Main from './components/main/main';
-import SignIn from './components/sign-in/sign-in';
-import MyList from './components/my-list/my-list';
-import FilmPage from './components/film-page/film-page';
-import AddReview from './components/add-review/add-review';
-import Player from './components/player/player';
-import NotFound from './components/not-found/not-found';
-import PrivateRoute from './components/private-route/private-route';
-import Sprite from './components/sprite/sprite';
-import Spinner from './components/spinner/Spinner';
-import ServerNotWorking from './components/server-not-working/server-not-working';
-import { getAppInitialized, getServerNotWorking } from './store/app/app-selector';
+import { AppRoute } from '../../constants';
+import { initializeApp } from '../../store/app/app-thunks';
+import { AsyncDispatch, State } from '../../store/store';
+import Main from '../main/main';
+import SignIn from '../sign-in/sign-in';
+import MyList from '../my-list/my-list';
+import FilmPage from '../film-page/film-page';
+import AddReview from '../add-review/add-review';
+import Player from '../player/player';
+import NotFound from '../not-found/not-found';
+import PrivateRoute from '../private-route/private-route';
+import Sprite from '../sprite/sprite';
+import Spinner from '../spinner/Spinner';
+import ServerNotWorking from '../server-not-working/server-not-working';
+import { getAppInitialized, getServerNotWorking } from '../../store/app/app-selector';
 
 const mapStateToProps = (state: State) => ({
   appInitialized: getAppInitialized(state),
