+++ project/src/index.tsx
@@ -2,7 +2,7 @@ import './css/main.css';
 import 'react-toastify/dist/ReactToastify.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import App from './components/app/app';
 import { Provider } from 'react-redux';
 import { store } from './store/store';
 import { BrowserRouter } from 'react-router-dom';
